trustguard_full/requirements.txt
aiogram==2.25.2
Flask==2.3.3
xrpl-py==2.1.1
